requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Document parsing
PyMuPDF==1.23.8
python-docx==0.8.11

# Machine Learning & NLP
transformers==4.35.2
sentence-transformers==2.2.2
torch==2.1.1
numpy==1.24.3

# Vector search
faiss-cpu==1.7.4

# Quantization support
bitsandbytes==0.41.2

# Utility
python-multipart==0.0.6
Pillow==10.0.1
aiofiles==23.2.1

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
